NF-4: Refactor thunk setup
