+++ src/store/rootReducer.ts
@@ -4,7 +4,7 @@ import { exampleThunk } from './slices';
 
 const rootReducer = combineReducers({
   exampleSaga: exampleSaga.reducer,
-  exampleThunk: exampleThunk.reducer
+  exampleThunk: exampleThunk.exampleThunkReducer
 });
 
 export default rootReducer;
