--- src/store/slices/example-thunk/selector.ts
@@ -1,3 +1,0 @@
-import { type RootState } from '@/store';
-
-export const selectNumber = (state: RootState) => state.exampleThunk.number;
